Reorder logging and add continue from log file.
The switchup in logging means that there is now nromally 1 extra entry for "labelled" compared to "pool_proba" and "train_proba". This will surely break the analysis package, but this makes much more sense, especially in an 'oracle' setting. Stopping after training is still supported, and should work.

We are now able to continue a simulation/review session starting from a log file. The settings of the previous session will be copied.

Bug fix: log entries (query id's) should be strings not integers.