Add HDF5 as a storage container for logging.
It works parallel to the still available JSON logger. It has an advantage in both speed and disk requirements (~x2.4).

The JSON logger was also updated, and it's version updated to 2.0, making it incompatible with log files created with an earlier version of ASReview. It will automatically detect if this is the case.