changed icon for irrelevant records
I found the X confusing because it is typically used for closing something. Therefore, I proposed a thumb-down icon. I also propose to change the title of the pop-up screen from 'details' to 'overview'. This because clicking the button brings you back to the overview page and not not more details about the record.