Run_model only trains and save last_training set
Removed all labeling actions from run_model.py. This should only train a new model and save the probabilities in the state. The training set is saved in the state at the same time, since later it will not be available. Next step should be to link the labeling in the api directly to the state. Right now this gives errors.