Adds state functions for queue mechanism
I added all functions I think are necessary for implementing a pool queueing mechanism in the state. After querying, a record is moved from pool to pending. After labeling a record is moved from pending to labeled. The functions add_labeling_data, get_pool_labeled_pending and query_top_ranked still require testing. Then the state should be ready to be connected to the reviewer and the frontend.