Add new reviewer
I finished the new version of the reviewer, updated the tests, the query strategies and balance models. Connected the frontend functions to the reviewer/state for a big part. Right now all test work, except those related to the 'triple' balance strategy and the 'mixed' query strategy. Frontend setup works until the training of the model.