Connect webapp to new state/reviewer
The functions in webapp are now connected to the functions in the new state/reviewer. Frontend seems to be working more or less: After model training you need to reload the page. The triple balance strategy and mixed query strategies do not work yet. Also, there are multiple functions in the frontend for labeling an item. There should be one function for labeling and one for updating a record.