Fix record_id issues
The ASReviewData class now also uses the standard record_ids instead of the record_ids from the dataset. However, the record_id column in the original dataset might get overwritten at export.

Fixed some bugs with the last_ranking table in the converter. Made a quick fix for exporting to excel, frontend uses "excel", but backend "xlsx". Also stopped the deletion of the '__MACOSX' file, since that apparently causes errors.

After this commit the project deletion still does not seem to be working and I've disabled the test for now. Also there still seems to be a bug regarding the order in which the records are shown.