Update simulation_results.rst
I would suggest to add a mention of the performance metrics that the ASReview Insights package provides in the ReadtheDocs of the ASReview software, so that users are immediately pointed to the ASReview Insights page when looking for these metrics on the RTD page. 

Motivation: I was looking for the documentation on performance metrics on the RTD page and couldn't find it, because it's a separate package that's not part of the core ASReview code. Simply adding in the search terms in the "Analyzing Results" page on simulations via CLI will help direct users to the right location more easily, as they will turn up as results when using the search bar.