Make SQL INSERT column specific
This makes all SQL 'INSERT' statements specify the columns into which data is inserted, even if data is inserted to all columns. This means that if an extra column is added to a table in a future version, the existing inserts will not give errors. Also added a test that adds a column to the results and last_ranking table and then tests some state functionality.