Remove User from ASReviewProject
This removes the User and Project database models from the ASReviewProject class. In asreview.project there is now only a function that returns a list of initialized project from a list of filepaths. Also, there is no more need to dynamically add an owner_id attribute to a ASReviewProject instance, because this was used only once as far as I could tell. In webapp/project.py all database related things are handled.