Trim .dockerignore to prevent dirty version in Docker
Versioneer uses git to determine the package version, and .dockerignore
will prevent unnecessary files from being sent to the build stage.
Since git is being used and the files are not there, they appear as
"deleted" under git, which means that the repo is dirty.
The solution is to not remove these files with .dockerignore.