auth_tool: allow adding users with non-standard origin
Currently `asreview auth-tool` assumes users added using the `--json` option have `asreview` as their `origin` attribute. That means users must have a valid email and a password.

However, for remote authentication mechanisms, we might want to create users that do not meet this criteria. In particular, we may want to pre-add valid users to the database when using Remote User auth, so that valid users are already visible in the app *before* they are created when they first login.